+++ packages/fs-detectors/src/index.ts
@@ -9,7 +9,7 @@ export {
   type DetectServicesOptions,
   type DetectServicesResult,
   type ResolvedService,
-} from './detect-services';
+} from './services';
 export { detectFileSystemAPI } from './detect-file-system-api';
 export {
   detectFramework,
